[type] Add support for 'unsigned'
Since the support actually depends on the dialect, the String() function
now takes the dialect as an argument.

For dialects that does not support unsigned, the int types are upgraded
to bigger ones so unsigned values can fit in. This behavior was chosen
out of consistency with what Mapper does.

Related to #80