Refactor primary key and auto increment support
The primary key can now be set directly on the column, and auto
increment columns DDL is generated by the dialect itself.

This small refactoring allow auto increment to work with postgres and
sqlite.

But it calls for a bigger refactoring where the dialects are in charge
of generating the sql code (throught a Compiler interface most
probably).

Closes #85